server/pyproject.toml
[project]
name = "didwebvh-server-py"
version = "0.3.4"
description = "DID WebVH Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.7.0",
    "aries-askar>=0.3.2",
    "bases==0.3.0",
    "cached-property>=1.5.2",
    "canonicaljson>=2.0.0",
    "click>=8.1.7",
    "fastapi>=0.115.2",
    "h11>=0.14.0",
    "idna>=3.7",
    "inflection>=0.5.1",
    "multiformats==0.3.1.post4",
    "multiformats-config>=0.3.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "pydantic-core>=2.20.1",
    "python-dotenv>=1.0.1",
    "sniffio>=1.3.1",
    "starlette>=0.38.6",
    "typing-validation==1.2.11.post4",
    "typing-extensions>=4.12.2",
    "uvicorn==0.29.0",
    "validators>=0.34.0",
    "jsonlines>=4.0.0",
    "pytest-asyncio>=0.25.2",
    "jcs>=0.2.1",
    "anoncreds>=0.2.0",
    "pytest-dependency>=0.6.0",
    "did-webvh>=1.0.0",
    "requests>=2.32.4",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "jinja2-eval>=0.1.2",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.0",
    "psycopg2-binary>=2.9.11",
]

[project.urls]
Repository = "https://github.com/decentralized-identity/didwebvh-server-py"

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

[tool.ruff]
lint.select = ["B006", "C", "D", "E", "F"]

# https://www.pydocstyle.org/en/2.1.1/error_codes.html
# https://docs.astral.sh/ruff/rules/#pydocstyle-d
lint.ignore = [
    # Google Python Doc Style
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D401",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
    "D202", # Allow blank line after docstring
    "D104", # Don't require docstring in public package
]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["B006", "D", "E501", "F841"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
